:zap: refactor: 将子弹碰撞独立
