+++ Assets/Prefabs/冲击波/Death_RemoteAttack.prefab
@@ -13,7 +13,7 @@ GameObject:
   - component: {fileID: 4758487640691272758}
   - component: {fileID: 3413898250392693238}
   - component: {fileID: 8331121822491177537}
-  - component: {fileID: 5389341420281520918}
+  - component: {fileID: 8544406023448721482}
   m_Layer: 3
   m_Name: Death_RemoteAttack
   m_TagString: EnemyBullet
@@ -145,10 +145,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: de7cbb15cdfc7404c80eef2c3cb95808, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speed: 10
+  speed: 20
   liveTime: 5
   facing: 1
---- !u!114 &5389341420281520918
+--- !u!114 &8544406023448721482
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -157,7 +157,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1607253366291472046}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6f6d18e1f3eca4e46af9fdb80fe9db40, type: 3}
+  m_Script: {fileID: 11500000, guid: 3a0f59c9bade175488a2ea9df4ddfef2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  attackDamage: 2
+  damage: 1
