+++ Assets/Scripts/ActionGame/RemoteAttack/GirlHeroBulletPhysicsComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21ea15f4b742dad4b83991fd953c03b1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
